backend/requirements.txt
fastapi==0.115.8
uvicorn==0.22.0
PyJWT==2.10.1
bcrypt==4.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.2.1
pydantic==2.6.3
Pillow==10.2.0
RPi.GPIO==0.7.1
picamera==1.13